requirements.txt
anthropic>=0.40.0
openai>=1.50.0
google-generativeai>=0.8.0
mistralai>=1.2.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
matplotlib>=3.8.0
seaborn>=0.13.0
python-dotenv>=1.0.0
tqdm>=4.66.0
